# Capture fewer variables to avoid dynamic memory allocation