--- a/TARGETS
+++ b/TARGETS
@@ -727,13 +727,16 @@ rocks_cpp_library_wrapper(name="rocksdb_stress_lib", srcs=[
     ], headers=None)
 
 
+rocks_cpp_library_wrapper(name="rocksdb_bench_lib", srcs=["microbench/bench_util.cc"], headers=None)
+
+
 cpp_binary_wrapper(name="db_stress", srcs=["db_stress_tool/db_stress.cc"], deps=[":rocksdb_stress_lib"], extra_preprocessor_flags=[], extra_bench_libs=False)
 
-cpp_binary_wrapper(name="ribbon_bench", srcs=["microbench/ribbon_bench.cc"], deps=[], extra_preprocessor_flags=[], extra_bench_libs=True)
+cpp_binary_wrapper(name="ribbon_bench", srcs=["microbench/ribbon_bench.cc"], deps=[":rocksdb_bench_lib"], extra_preprocessor_flags=[], extra_bench_libs=True)
 
-cpp_binary_wrapper(name="db_basic_bench", srcs=["microbench/db_basic_bench.cc"], deps=[], extra_preprocessor_flags=[], extra_bench_libs=True)
+cpp_binary_wrapper(name="db_basic_bench", srcs=["microbench/db_basic_bench.cc"], deps=[":rocksdb_bench_lib"], extra_preprocessor_flags=[], extra_bench_libs=True)
 
-cpp_binary_wrapper(name="reader_bench", srcs=["microbench/reader_bench.cc"], deps=[], extra_preprocessor_flags=[], extra_bench_libs=True)
+cpp_binary_wrapper(name="reader_bench", srcs=["microbench/reader_bench.cc"], deps=[":rocksdb_bench_lib"], extra_preprocessor_flags=[], extra_bench_libs=True)
 
 add_c_test_wrapper()