--- a/memtable/memtablerep_bench.cc
+++ b/memtable/memtablerep_bench.cc
@@ -552,7 +552,6 @@ class ReadWriteBenchmark : public Benchmark {
     }
   }
 };
-
 }  // namespace ROCKSDB_NAMESPACE
 
 void PrintWarnings() {