--- a/microbench/CMakeLists.txt
+++ b/microbench/CMakeLists.txt
@@ -4,7 +4,7 @@ find_package(Threads REQUIRED)
 file(GLOB_RECURSE ALL_BENCH_CPP *.cc)
 foreach(ONE_BENCH_CPP ${ALL_BENCH_CPP})
   get_filename_component(TARGET_NAME ${ONE_BENCH_CPP} NAME_WE)
-  add_executable(${TARGET_NAME} ${ONE_BENCH_CPP})
+  add_executable(${TARGET_NAME} ${ONE_BENCH_CPP} bench_util.cc bench_util.h)
   target_link_libraries(${TARGET_NAME} ${ROCKSDB_LIB} benchmark::benchmark
           ${CMAKE_THREAD_LIBS_INIT})
   # run benchmark like a test, if added, the benchmark tests could be run by `ctest -R Bench_`