--- a/table/block_based/block_like_traits.h
+++ b/table/block_based/block_like_traits.h
@@ -20,26 +20,6 @@ class BlocklikeTraits;
 template <typename T, CacheEntryRole R>
 Cache::CacheItemHelper* GetCacheItemHelperForRole();
 
-template <typename TBlocklike>
-Cache::CreateCallback GetCreateCallback(size_t read_amp_bytes_per_bit,
-                                        Statistics* statistics, bool using_zstd,
-                                        const FilterPolicy* filter_policy) {
-  return [read_amp_bytes_per_bit, statistics, using_zstd, filter_policy](
-             const void* buf, size_t size, void** out_obj,
-             size_t* charge) -> Status {
-    assert(buf != nullptr);
-    std::unique_ptr<char[]> buf_data(new char[size]());
-    memcpy(buf_data.get(), buf, size);
-    BlockContents bc = BlockContents(std::move(buf_data), size);
-    TBlocklike* ucd_ptr = BlocklikeTraits<TBlocklike>::Create(
-        std::move(bc), read_amp_bytes_per_bit, statistics, using_zstd,
-        filter_policy);
-    *out_obj = reinterpret_cast<void*>(ucd_ptr);
-    *charge = size;
-    return Status::OK();
-  };
-}
-
 template <>
 class BlocklikeTraits<BlockContents> {
  public: