#10453 uses the memory from stack instead of heap, but the user might save the
callback and invoke that later, if the read is return, it causes ASAN
stack-use-after-return error.  So we have to switch back to include captured
local variable to `std::function`, so at least when the user copy the callback,
it also copies the captured variables. Change it to capture the reference of the
table reader `BlockBasedTable`

Test Plan: Added benchmark for TableReader.Get(), which shows ~2.5% improvement:
```
./reader_bench --benchmark_filter=BlockBasedTableGet --benchmark_repetitions=10
# before the change:
BlockBasedTableGet/iterations:1000000_mean          741 ns          741 ns           10
# after the change:
BlockBasedTableGet/iterations:1000000_mean          722 ns          722 ns           10
```