@pdillinger 2022-11-09
By the way, I have a refactoring in progress that should completely solve this issue