@pdillinger 2022-11-10
> By the way, I have a refactoring in progress that should completely solve this issue

I should say my solution also depends on the table reader staying alive, but *should be* cleaner because of other refactoring. That's the hope.